requirements.txt
pillow==10.2.0
pystray==0.19.5
opencv-python==4.9.0.80
onnxruntime-gpu==1.16.3
insightface==0.7.3